Implemented new subscribing mechanizm.
Applied the new mechanizm to MetricLoader, thus code became simpler,
and smaller.
It allows subscribing to various events on MetricLoader, currently only
the "doneFetching" event is implemented, but it is very simple to add
new ones.
Rewrote the widgets as well to use this new mechanizm instead of
the old one. This mechanizm will be used on the cloneviewer side too.